+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 New functionality added in a backwards-compatible manner:
 * [OLMIS-6350](https://openlmis.atlassian.net/browse/OLMIS-6350): Added openlmisOkIcon component.
 * [OLMIS-6380](https://openlmis.atlassian.net/browse/OLMIS-6380): Added openlmis-datetime module.
-* [OLMIS-6361](https://openlmis.atlassian.net/browse/OLMIS-6361): Added openlmis-message component.
+* [OLMIS-6361](https://openlmis.atlassian.net/browse/OLMIS-6361): Added openlmis-message and openlmis-message-creation-details components.
 * [OLMIS-6359](https://openlmis.atlassian.net/browse/OLMIS-6359): Added missing message openlmisForm.maxlength.
 
 Improvements: